Move ParallelTestRunner to its own file, clean up lots of prints
